move template printers to genericclioptions
