+++ pkg/kubectl/genericclioptions/template_flags.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package printers
+package genericclioptions
 
 import (
 	"fmt"
@@ -23,30 +23,32 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"k8s.io/kubernetes/pkg/kubectl/genericclioptions"
+	"k8s.io/kubernetes/pkg/kubectl/genericclioptions/printers"
 )
 
 // templates are logically optional for specifying a format.
 // this allows a user to specify a template format value
-// as --output=jsonpath=
-var jsonFormats = map[string]bool{
-	"jsonpath":      true,
-	"jsonpath-file": true,
+// as --output=go-template=
+var templateFormats = map[string]bool{
+	"template":         true,
+	"go-template":      true,
+	"go-template-file": true,
+	"templatefile":     true,
 }
 
-// JSONPathPrintFlags provides default flags necessary for template printing.
+// GoTemplatePrintFlags provides default flags necessary for template printing.
 // Given the following flag values, a printer can be requested that knows
 // how to handle printing based on these values.
-type JSONPathPrintFlags struct {
+type GoTemplatePrintFlags struct {
 	// indicates if it is OK to ignore missing keys for rendering
 	// an output template.
 	AllowMissingKeys *bool
 	TemplateArgument *string
 }
 
-func (f *JSONPathPrintFlags) AllowedFormats() []string {
-	formats := make([]string, 0, len(jsonFormats))
-	for format := range jsonFormats {
+func (f *GoTemplatePrintFlags) AllowedFormats() []string {
+	formats := make([]string, 0, len(templateFormats))
+	for format := range templateFormats {
 		formats = append(formats, format)
 	}
 	return formats
@@ -55,15 +57,15 @@ func (f *JSONPathPrintFlags) AllowedFormats() []string {
 // ToPrinter receives an templateFormat and returns a printer capable of
 // handling --template format printing.
 // Returns false if the specified templateFormat does not match a template format.
-func (f *JSONPathPrintFlags) ToPrinter(templateFormat string) (ResourcePrinter, error) {
+func (f *GoTemplatePrintFlags) ToPrinter(templateFormat string) (printers.ResourcePrinter, error) {
 	if (f.TemplateArgument == nil || len(*f.TemplateArgument) == 0) && len(templateFormat) == 0 {
-		return nil, genericclioptions.NoCompatiblePrinterError{Options: f, OutputFormat: &templateFormat}
+		return nil, NoCompatiblePrinterError{Options: f, OutputFormat: &templateFormat}
 	}
 
 	templateValue := ""
 
 	if f.TemplateArgument == nil || len(*f.TemplateArgument) == 0 {
-		for format := range jsonFormats {
+		for format := range templateFormats {
 			format = format + "="
 			if strings.HasPrefix(templateFormat, format) {
 				templateValue = templateFormat[len(format):]
@@ -75,15 +77,15 @@ func (f *JSONPathPrintFlags) ToPrinter(templateFormat string) (ResourcePrinter, 
 		templateValue = *f.TemplateArgument
 	}
 
-	if _, supportedFormat := jsonFormats[templateFormat]; !supportedFormat {
-		return nil, genericclioptions.NoCompatiblePrinterError{OutputFormat: &templateFormat, AllowedFormats: f.AllowedFormats()}
+	if _, supportedFormat := templateFormats[templateFormat]; !supportedFormat {
+		return nil, NoCompatiblePrinterError{OutputFormat: &templateFormat, AllowedFormats: f.AllowedFormats()}
 	}
 
 	if len(templateValue) == 0 {
 		return nil, fmt.Errorf("template format specified but no template given")
 	}
 
-	if templateFormat == "jsonpath-file" {
+	if templateFormat == "templatefile" || templateFormat == "go-template-file" {
 		data, err := ioutil.ReadFile(templateValue)
 		if err != nil {
 			return nil, fmt.Errorf("error reading --template %s, %v\n", templateValue, err)
@@ -92,9 +94,9 @@ func (f *JSONPathPrintFlags) ToPrinter(templateFormat string) (ResourcePrinter, 
 		templateValue = string(data)
 	}
 
-	p, err := NewJSONPathPrinter(templateValue)
+	p, err := printers.NewGoTemplatePrinter([]byte(templateValue))
 	if err != nil {
-		return nil, fmt.Errorf("error parsing jsonpath %s, %v\n", templateValue, err)
+		return nil, fmt.Errorf("error parsing template %s, %v\n", templateValue, err)
 	}
 
 	allowMissingKeys := true
@@ -108,9 +110,9 @@ func (f *JSONPathPrintFlags) ToPrinter(templateFormat string) (ResourcePrinter, 
 
 // AddFlags receives a *cobra.Command reference and binds
 // flags related to template printing to it
-func (f *JSONPathPrintFlags) AddFlags(c *cobra.Command) {
+func (f *GoTemplatePrintFlags) AddFlags(c *cobra.Command) {
 	if f.TemplateArgument != nil {
-		c.Flags().StringVar(f.TemplateArgument, "template", *f.TemplateArgument, "Template string or path to template file to use when --output=jsonpath, --output=jsonpath-file.")
+		c.Flags().StringVar(f.TemplateArgument, "template", *f.TemplateArgument, "Template string or path to template file to use when -o=go-template, -o=go-template-file. The template format is golang templates [http://golang.org/pkg/text/template/#pkg-overview].")
 		c.MarkFlagFilename("template")
 	}
 	if f.AllowMissingKeys != nil {
@@ -118,11 +120,14 @@ func (f *JSONPathPrintFlags) AddFlags(c *cobra.Command) {
 	}
 }
 
-// NewJSONPathPrintFlags returns flags associated with
+// NewGoTemplatePrintFlags returns flags associated with
 // --template printing, with default values set.
-func NewJSONPathPrintFlags(templateValue string, allowMissingKeys bool) *JSONPathPrintFlags {
-	return &JSONPathPrintFlags{
-		TemplateArgument: &templateValue,
-		AllowMissingKeys: &allowMissingKeys,
+func NewGoTemplatePrintFlags() *GoTemplatePrintFlags {
+	allowMissingKeysPtr := true
+	templateValuePtr := ""
+
+	return &GoTemplatePrintFlags{
+		TemplateArgument: &templateValuePtr,
+		AllowMissingKeys: &allowMissingKeysPtr,
 	}
 }
